Update the accumulator and the main file
