+++ src/lib.rs
@@ -1,0 +1,5 @@
+pub mod bb_params;
+pub mod credential;
+pub mod commitment;
+pub mod witness;
+pub mod acumulator;
